# Conflicts:
#	Assets/1)Test Meccaniche/Scenes/CharacterTesteScene.unity
#	Assets/1)Test Meccaniche/Scripts/Interactables/Key/Interactable.cs.meta
#	Assets/1)Test Meccaniche/Scripts/Player/PlayerController.cs
#	Assets/1)Test Meccaniche/Scripts/Weapons/ShootComponent.cs

+++ Assets/1)Test Meccaniche/Scripts/PowerUps/ScriptableObject/PermanentPUSO.cs
@@ -8,8 +8,8 @@ namespace TheChroniclesOfEllen
     [CreateAssetMenu(menuName = "Window/Permanent", fileName = "Permanent")]
     public class PermanentPUSO : PowerUpSO
     {
-        public int damageIncrease;
-        public int healthIncrease;
+        public float damageIncrease;
+        public float healthIncrease;
 
         public override void SetPowerUpType()
         {
